requirements.txt
slack_bolt
requests
PyYAML
